Fix some casting/deprecation warnings, reset can keep state config

+++ Classes/DiskDriveService.h
@@ -32,7 +32,7 @@
 /**
  * Inserts the specified adf into the specified drive (df[0], df[1] ...).
  */
-- (void)insertDisk:(NSString *)adfPath intoDrive:(int)driveNumber;
+- (void)insertDisk:(NSString *)adfPath intoDrive:(NSUInteger)driveNumber;
 
 /**
  * Inserts the specified adfs into the drives corresponding to the adfs' positions in the array: adf at position 0 -> df0, adf at position 1 -> df1 ...
